pyproject.toml
[project]
name = "strands-otel-tracing-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.74",
    "botocore>=1.42.74",
    "langsmith>=0.7.22",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "strands-agents==1.32.0",
    "strands-agents-tools==0.2.23",
]
